Rename the HeroPoll cog to Dota.
Also move the heropoll command implementation to its own file.
